Moves font size functions into own folder
