+++ packages/lms/src/lib/utils/settings/fontSize/constants.ts
@@ -1,0 +1,2 @@
+export const tailwindFontSizes: string[] = ['text-sm', 'text-base', 'text-lg', 'text-xl', 'text-2xl'];
+export const proseFontSizes: string[] = ['prose-sm', 'prose-base', 'prose-lg', 'prose-xl', 'prose-2xl'];
