+++ packages/lms/src/lib/utils/settings/fontSize/index.ts
@@ -1,0 +1,6 @@
+export * from './constants';
+export * from './increase';
+export * from './decrease';
+export * from './reset';
+
+// import * as fontSize from "../fontSize"
