+++ packages/lms/src/lib/utils/storage/index.ts
@@ -3,4 +3,4 @@ export * from './remove';
 export * from './save';
 export * from './update';
 
-// import * as storage from "../storage"
+// import * as storage from "../fontSize"
